Skip first char
This ensures we don't count the first char as an occurence, leading to
incorrect highlighting.

For example, pressing 'f' at the start of `t t` would have highlighted
the second 't' in the secondary colour, since it is the second
occcurence of the char. Now, since we skip the first 't' the second 't'
is counted as the first occurence, leading to primary highlighting.